online review comments.

+++ engine/table/src/main/java/io/deephaven/engine/table/impl/sources/flat/FlatByteArraySource.java
@@ -67,7 +67,7 @@ public class FlatByteArraySource extends AbstractDeferredGroupingColumnSource<By
 
     @Override
     public final void set(long key, byte value) {
-        data[Math.toIntExact(key)] = value;
+        data[(int)key] = value;
     }
 
     @Override
@@ -93,8 +93,8 @@ public class FlatByteArraySource extends AbstractDeferredGroupingColumnSource<By
         final WritableByteChunk<? super Values> asByteChunk = destination.asWritableByteChunk();
         final MutableInt destPosition = new MutableInt(0);
         rowSequence.forAllRowKeyRanges((long start, long end) -> {
-            final int rangeLength = Math.toIntExact(end - start + 1);
-            asByteChunk.copyFromTypedArray(data, Math.toIntExact(start), destPosition.getAndAdd(rangeLength), rangeLength);
+            final int rangeLength = (int)(end - start + 1);
+            asByteChunk.copyFromTypedArray(data, (int)start, destPosition.getAndAdd(rangeLength), rangeLength);
         });
         asByteChunk.setSize(destPosition.intValue());
     }
@@ -126,9 +126,9 @@ public class FlatByteArraySource extends AbstractDeferredGroupingColumnSource<By
 
     @Override
     public long resetWritableChunkToBackingStoreSlice(@NotNull ResettableWritableChunk<?> chunk, long position) {
-        final int capacity = Math.toIntExact(data.length - position);
+        final int capacity = (int)(data.length - position);
         ResettableWritableByteChunk resettableWritableByteChunk = chunk.asResettableWritableByteChunk();
-        resettableWritableByteChunk.resetFromTypedArray(data, Math.toIntExact(position), capacity);
+        resettableWritableByteChunk.resetFromTypedArray(data, (int)position, capacity);
         return capacity;
     }
 
@@ -154,9 +154,9 @@ public class FlatByteArraySource extends AbstractDeferredGroupingColumnSource<By
 
     @Override
     public Chunk<? extends Values> getChunk(@NotNull GetContext context, long firstKey, long lastKey) {
-        final int len = Math.toIntExact(lastKey - firstKey + 1);
+        final int len = (int)(lastKey - firstKey + 1);
         final GetContextWithResettable contextWithResettable = (GetContextWithResettable) context;
-        return contextWithResettable.resettableByteChunk.resetFromTypedArray(data, Math.toIntExact(firstKey), len);
+        return contextWithResettable.resettableByteChunk.resetFromTypedArray(data, (int)firstKey, len);
     }
 
     @Override
@@ -178,8 +178,8 @@ public class FlatByteArraySource extends AbstractDeferredGroupingColumnSource<By
         final ByteChunk<? extends Values> asByteChunk = src.asByteChunk();
         final MutableInt srcPos = new MutableInt(0);
         rowSequence.forAllRowKeyRanges((long start, long end) -> {
-            final int rangeLength = Math.toIntExact(end - start + 1);
-            asByteChunk.copyToTypedArray(srcPos.getAndAdd(rangeLength), data, Math.toIntExact(start), rangeLength);
+            final int rangeLength = (int)(end - start + 1);
+            asByteChunk.copyToTypedArray(srcPos.getAndAdd(rangeLength), data, (int)start, rangeLength);
         });
     }
 
@@ -199,7 +199,7 @@ public class FlatByteArraySource extends AbstractDeferredGroupingColumnSource<By
             if (longKey == RowSet.NULL_ROW_KEY) {
                 byteDest.set(ii, NULL_BYTE);
             } else {
-                final int key = Math.toIntExact(longKey);
+                final int key = (int)longKey;
                 byteDest.set(ii, getUnsafe(key));
             }
         }
